// old/src/lib/types.ts
export interface ProductPhotos {
  id: number;
  name: string;
  product_id: number;
  defaultphoto: number;
  path: string;
  status: string;
  media_type?: string;
  file_size?: number;
  asset_url: string;
}